Rename OpcodeTables to InstructionSet
