+++ DameDis/DameDis.vcxproj
@@ -155,7 +155,7 @@
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">NotUsing</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">NotUsing</PrecompiledHeader>
     </ClCompile>
-    <ClCompile Include="OpcodeTables.h" />
+    <ClCompile Include="InstructionSet.h" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Cartridge.h" />
